requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
azure-core==1.34.0
azure-storage-blob==12.25.1
fastapi==0.115.12
flatbuffers==25.2.10
fonttools==4.58.0
h5py==3.13.0
keras==2.14.0
numpy==1.26.4
packaging==25.0
protobuf==4.25.7
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
scipy==1.15.3
six==1.17.0
SQLAlchemy==2.0.41
starlette==0.46.2
tensorflow==2.14.1
torch==1.13.1
typing_extensions==4.13.2
tzlocal==5.3.1
uvicorn==0.34.2
